requirements.txt
mysql_connector_repackaged==0.3.1
numpy==1.26.4
Pillow==10.2.0
streamlit==1.32.2
PyMysql
mysql-connector
py-cpuinfo
torch
torchvision
opencv-python
random
math